==> 1 Data prep_reading in data.Rmd <==
---
title: "2019-11-09 data prep"
author: "Alexander Klueber"
date: "11/9/2019"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE,
                      message = FALSE,
                      error = FALSE,
                      warning = FALSE)
```
```{r}
library(readxl)
library(tibble)
library(lubridate)
library(tidyverse)
```

```{r, include =FALSE}
# A. CDP data

# A.1 listing the data paths for CDP data in all available years
# Challenge: each XLSX file had a plethora of sheets (~50) which differed in structure. Preliminary work therefore included selecting the relevant sheets and saving them in a separate folder as new documents

xl_2013 <- "C:/Users/Alexander Klueber/Desktop/University/Coursework/Fall 2019/GOV 1005 Data/Final project/Final project/Carbon footprint/2013_carbon footprint.xlsx"
xl_2014 <- "C:/Users/Alexander Klueber/Desktop/University/Coursework/Fall 2019/GOV 1005 Data/Final project/Final project/Carbon footprint/2014_carbon footprint.xlsx"
xl_2015 <- "C:/Users/Alexander Klueber/Desktop/University/Coursework/Fall 2019/GOV 1005 Data/Final project/Final project/Carbon footprint/2015_carbon footprint.xlsx"
xl_2016 <- "C:/Users/Alexander Klueber/Desktop/University/Coursework/Fall 2019/GOV 1005 Data/Final project/Final project/Carbon footprint/2016_carbon footprint.xlsx"
xl_2017 <- "C:/Users/Alexander Klueber/Desktop/University/Coursework/Fall 2019/GOV 1005 Data/Final project/Final project/Carbon footprint/2017_carbon footprint.xlsx"
xl_2018 <- "C:/Users/Alexander Klueber/Desktop/University/Coursework/Fall 2019/GOV 1005 Data/Final project/Final project/Carbon footprint/2018_carbon footprint.xlsx"

# A.2 Defining functions to load in the data (note: different functions per year because layout of Excels and naming conventions change between years) 
  # Creating/Selecting same lines for all years and attaching identical label ():
    #1. Organisation
    #2. Account_Number
    #3. Country
    #4. Industry
    #5. Ticker
    #6. ISIN
    #7. Reporting_year
    #8. Scope_1
    #9. Scope_2
    #10. Bio_CO2
    #11. "Year_acc_id" (Combination of Reporting_year & " / " & Accounting_Number to create unique ID)
    #12. Publication_data (Year of reporting)


  #Notes: 
    #Excluded 2011 & 2012, because didn't have Ticker and ISIN for accounts
    # Translated Reporting_from and Reporting_to to an accounting year for 2013 - 2016. Year based on where majority of accounting occured.
    # Where market and location based scope 2 emissions were available, preference was given to location based emissions
    # From 2016 on combination of two tables necessary, because scope 1 & bio, scope 2 emissions on separate tabs
    # From 2018 on combination of three tables necessary, because scope 1, scope 2, and bio emissions on separate tabs
 
  load_data_2013 <- function(input){
    ccf_1 <- read_xlsx(input) %>%
             rename(Reporting_from = "Reporting Period\r\nFrom", #Rename columns that include dates
                    Reporting_to = "Reporting Period\r\nTo") %>%  
             mutate(Reporting_from_2 = as.Date(Reporting_from, format ="%d %b %Y"), # Cast date columns in date format and determine accounting year with if or logic based on Reporting To Month
                    Reporting_to_2 = as.Date(Reporting_to, format ="%d %b %Y"),
                    Reporting_year = ifelse(month(Reporting_to_2) > 6, 
                                            year(Reporting_to_2), 
                                            year(Reporting_from_2))) %>%
             select(Organisation = "Organisation", #Select variables from data frame based on names
             Account_Number = "Account_No", 
             Country = "Country", 
             Industry = "GICS Sector", 
             Ticker = "Ticker", 
             ISIN = "ISIN", 
             Reporting_year, 
             Scope_1 = "8.2. \r\nPlease provide your gross global Scope 1 emissions figures in metric tonnes CO2e",                  Scope_2 = "8.3. \r\nPlease provide your gross global Scope 2 emissions figures in metric tonnes CO2e", 
             Bio_CO2 = "8.8a. \r\nPlease provide the emissions in metric tonnes CO2") %>%
             unite("Year_acc_id", c(Reporting_year, Account_Number), sep =" / ") %>% # Create unique ID per line
             mutate(Publication_data = 2013) # Add line for year of publication
    return(ccf_1)
  }
  
  
  load_data_2014 <- function(input){
    ccf_1 <- read_xlsx(input) %>%
             rename(Reporting_from = "Reporting Period\r\nFrom",
                    Reporting_to = "Reporting Period\r\nTo") %>%  
             mutate(Reporting_from_2 = as.Date(Reporting_from, format ="%d %b %Y"),
                    Reporting_to_2 = as.Date(Reporting_to, format ="%d %b %Y"),
                    Reporting_year = ifelse(month(Reporting_to_2) > 6, 
                                            year(Reporting_to_2), 
                                            year(Reporting_from_2))) %>%
             select(Organisation = "Organisation", 
             Account_Number = "Account_No", 
             Country = "Country", 
             Industry = "GICS Sector", 
             Ticker = "Ticker", 
             ISIN = "ISIN", 
             Reporting_year,
             Scope_1 = "CC8.2. Please provide your gross global Scope 1 emissions figures in metric tonnes CO2e",   
             Scope_2 = "CC8.3. Please provide your gross global Scope 2 emissions figures in metric tonnes CO2e", 
             Bio_CO2 = "CC8.9a. Please provide the emissions from biologically sequestered carbon in metric tonnes CO2")%>%
             unite("Year_acc_id", c(Reporting_year, Account_Number), sep =" / ")%>%
             mutate(Publication_data = 2014)
    return(ccf_1)
  }

  
  load_data_2015 <- function(input){
    ccf_1 <- read_xlsx(input) %>%
             rename(Reporting_from = "Reporting Period\r\nFrom",
                    Reporting_to = "Reporting Period\r\nTo") %>%  
             mutate(Reporting_from_2 = as.Date(Reporting_from, format ="%d %b %Y"),
                    Reporting_to_2 = as.Date(Reporting_to, format ="%d %b %Y"),
                    Reporting_year = ifelse(month(Reporting_to_2) > 6, 
                                            year(Reporting_to_2), 
                                            year(Reporting_from_2))) %>%
             select(Organisation = "Organisation", 
             Account_Number = "Account_No", 
             Country = "Country", 
             Industry = "GICS Sector", 
             Ticker = "Ticker", 
             ISIN = "ISIN", 
             Reporting_year, 
             Scope_1 = "CC8.2. Please provide your gross global Scope 1 emissions figures in metric tonnes CO2e",   
             Scope_2 = "CC8.3. Please provide your gross global Scope 2 emissions figures in metric tonnes CO2e", 
             Bio_CO2 = "CC8.9a. Please provide the emissions from biologically sequestered carbon in metric tonnes CO2")%>%
        unite("Year_acc_id", c(Reporting_year, Account_Number), sep =" / ")%>%
        mutate(Publication_data = 2015)
    return(ccf_1)
  }


# For 2016, combination of two sheets required, also for scope 2: market-based and location based available (methodology: always use location based, except when not available, then use market-based)
  load_data_2016 <- function(input){
    tab_names <- excel_sheets(input)
    corporate_carbon_footprints <- lapply(tab_names, function(x) read_excel(path = input, sheet = x))
    ccf_1 <- data.frame(corporate_carbon_footprints[1]) %>%
             rename(Reporting_from = "Reporting.Period..From",
                    Reporting_to = "Reporting.Period..To") %>%  
             mutate(Reporting_from_2 = as.Date(Reporting_from, format ="%d %b %Y"),
                    Reporting_to_2 = as.Date(Reporting_to, format ="%d %b %Y"),
                    Reporting_year = ifelse(month(Reporting_to_2) > 6, 
                                            year(Reporting_to_2), 
                                            year(Reporting_from_2))) %>%
             select(Organisation = "Organisation", 
             Account_Number = "Account.No", 
             Country = "Country", 
             Industry = "Industry.Group", 
             Ticker = "Ticker", 
             ISIN = "ISIN",
             Reporting_year,
             Scope_1 = "CC8.2..Please.provide.your.gross.global.Scope.1.emissions.figures.in.metric.tonnes.CO2e",
             Bio_CO2 = "CC8.9a..Please.provide.the.emissions.from.biologically.sequestered.carbon.relevant.to.your.organization.in.metric.tonnes.CO2")
    ccf_2 <- data.frame(corporate_carbon_footprints[2]) %>%
             select(Scope_2_location = "CC8.3a..Please.provide.your.gross.global.Scope.2.emissions.figures.in.metric.tonnes.CO2e......CC8.3a..Scope.2..location.based.U.200B.", 
             Scope_2_market = "CC8.3a..Scope.2..market.based..if.applicable..U.200B.") %>%
             mutate(Scope_2 = ifelse(!is.na(Scope_2_location),Scope_2_location,Scope_2_market))
    ccf <- data.frame(ccf_1, ccf_2) %>%
             select(Organisation, Account_Number, Country, Industry, Ticker, ISIN, Reporting_year, Scope_1, Scope_2, Bio_CO2)%>%
            unite("Year_acc_id", c(Reporting_year, Account_Number), sep =" / ")%>%
            mutate(Publication_data = 2016)
    return(ccf)
  }
  

# For 2017, combination of two sheets required, also for scope 2: market-based and location based available (methodology: always use location based, except when not available, then use market-based)

load_data_2017 <- function(input){
    tab_names <- excel_sheets(input)
    corporate_carbon_footprints <- lapply(tab_names, function(x) read_excel(path = input, sheet = x))
    ccf_1 <- data.frame(corporate_carbon_footprints[1]) %>%
             select(Organisation = "account_name", 
             Account_Number = "account_id", 
             Country, 
             Industry, 
             Ticker, 
             ISIN, 
             Reporting_year = "accounting_year", 
             Scope_1 = "CC8.2...Please.provide.your.gross.global.Scope.1.emissions.figures.in.metric.tonnes.CO2e",
             Bio_CO2 = "CC8.9a...Please.provide.the.emissions.from.biologically.sequestered.carbon.relevant.to.your.organization.in.metric.tonnes.CO2")
    ccf_2 <- data.frame(corporate_carbon_footprints[2]) %>%
             select(Scope_2_location = "CC8.3a.C1...Please.provide.your.gross.global.Scope.2.emissions.figures.in.metric.tonnes.CO2e....Scope.2..location.based.", 
             Scope_2_market = "CC8.3a.C2...Please.provide.your.gross.global.Scope.2.emissions.figures.in.metric.tonnes.CO2e....Scope.2..market.based..if.applicable..") %>%
             mutate(Scope_2 = ifelse(!is.na(Scope_2_location),Scope_2_location,Scope_2_market))
    ccf <- data.frame(ccf_1, ccf_2) %>%
             select(Organisation, Account_Number, Country, Industry, Ticker, ISIN, Reporting_year, Scope_1, Scope_2, Bio_CO2)%>%
           unite("Year_acc_id", c(Reporting_year, Account_Number), sep =" / ")%>%
           mutate(Publication_data = 2017)
    return(ccf)
  }
  

load_data_2018 <- function(input){
    tab_names <- excel_sheets(input)
    corporate_carbon_footprints <- lapply(tab_names, function(x) read_excel(path = input, sheet = x))
    ccf_1 <- data.frame(corporate_carbon_footprints[1]) %>%
             mutate(Reporting_year = 2017) %>%
             select(Organisation = "Organization", 
             Account_Number = "Account.number", 
             Country = "Country", 
             Industry = "Primary.sector", 
             Ticker = "Ticker", 
             ISIN = "ISIN", 
             Reporting_year, 
             Scope_1 = "C6.1_C1_What.were.your.organization.s.gross.global.Scope.1.emissions.in.metric.tons.CO2e....Gross.global.Scope.1.emissions..metric.tons.CO2e.")
    ccf_2 <- data.frame(corporate_carbon_footprints[2]) %>%
             select(Scope_2_location = "C6.3_C1_What.were.your.organization.s.gross.global.Scope.2.emissions.in.metric.tons.CO2e....Scope.2..location.based", 
             Scope_2_market = "C6.3_C2_What.were.your.organization.s.gross.global.Scope.2.emissions.in.metric.tons.CO2e....Scope.2..market.based..if.applicable.") %>%
             mutate(Scope_2 = ifelse(!is.na(Scope_2_location),Scope_2_location,Scope_2_market))
    ccf_3 <- data.frame(corporate_carbon_footprints[3]) %>%
             select(Bio_CO2 = "C6.7a_Provide.the.emissions.from.biologically.sequestered.carbon.relevant.to.your.organization.in.metric.tons.CO2.")
    ccf <- data.frame(ccf_1, ccf_2, ccf_3) %>%
             select(Organisation, Account_Number, Country, Industry, Ticker, ISIN, Reporting_year, Scope_1, Scope_2, Bio_CO2)%>%
           unite("Year_acc_id", c(Reporting_year, Account_Number), sep =" / ")%>%
           mutate(Publication_data = 2018)
    return(ccf)
}


# A.3 Applying functions to load in the data 
  
ccf_2013 <- load_data_2013(xl_2013)

ccf_2014 <- load_data_2014(xl_2014)

ccf_2015 <- load_data_2015(xl_2015)

ccf_2016 <- load_data_2016(xl_2016)

ccf_2017 <- load_data_2017(xl_2017)

ccf_2018 <- load_data_2018(xl_2018)


# A.4 Uniting dataframes per year into one dataframe  

ccf_total <- as.data.frame(rbind(ccf_2018, ccf_2017, ccf_2016, ccf_2015, ccf_2014, ccf_2013))


# A.5 To get most up to date data for every account for every available year: group by Year_acc_id and then slice for first to get newest data if there are multiple, separate Year_acc_id again  

ccf_total_n <- ccf_total %>%
  group_by(Year_acc_id) %>%
  slice(1) %>%
  separate(Year_acc_id, c("Year", "Account_Number"), sep = " / ")


# A.6 Cast CO2 scopes as integers 

ccf_total_n$Scope_1 <- as.integer(ccf_total_n$Scope_1)
ccf_total_n$Scope_2 <- as.integer(ccf_total_n$Scope_2)
ccf_total_n$Bio_CO2 <- as.integer(ccf_total_n$Bio_CO2)


# A.7 Create sum column for Total 

ccf_total_n_a <- ccf_total_n %>%
  rowwise() %>% 
  mutate(Total = sum(Scope_1,Scope_2,Bio_CO2, na.rm = TRUE))


# A.8 Create new directory ("clean-data") to save data and safe consolidate carbon footprint table there

dir.create("clean-data")
write_rds(ccf_total_n_a, "clean-data/ccf_combined.rds", compress = "none")

```
```{r}
#B. Bloomberg data on market capitalization of companies


# B.1 Adding Bloomberg data: Download list of ISIN's to determine market value per 09/30 via Bloomberg tunnels (Bloomberg only has Excel interface at disposal, available at Business School library) / Note: Market Cap according on main listed exchange acc. to Bloomberg data based on ISIN per 09/30/19 / Cast marketcap as number

Bloomberg_list <- ccf_total_n_a %>%
                  select("ISIN") %>%
                  distinct(ISIN)


# B.2 Write table into folder for download and utilitization in Bloomberg terminal

write_csv(Bloomberg_list, "clean-data/Bloomberg.cvs")


# B.3 Reading in Excel with Bloomberg terminal output at tap 2: data prep

```


```{r}
#C. SEC filings

#C.1 Load in filings

SEC_filings <- read_xlsx("C:/Users/Alexander Klueber/Desktop/University/Coursework/Fall 2019/GOV 1005 Data/Final project/Final project/Funds/HMC filed securities Q3 2019.xlsx")


# Challenge: Naming conventions between companies across data sets differ significantly. Same for stock symbols which might relate to different securities of the same company. Therefore, manual cleaning of initial data set to harmonize names/tickers based on the information that can be matched with the carbon footprint data set for all companies whose relative share exceeds 0.1%


#C.2 Based on (manually revised) Tickers and names, group companies based on tickers to get share of porfolio at last filing

SEC_filings_cons <- SEC_filings %>%
  group_by(Ticker) %>%
  summarize(`Share of portfolio`= sum(`Direct share of portfolio`))


#C.3 Write file to clean-data folder

write_rds(SEC_filings_cons, "clean-data/SEC_filings_cons.rds", compress = "none")
```

```{r}
#D. Endowment utilitization 


#D.1 Read in file of endowment split

Endowment_split <- read_xlsx("C:/Users/Alexander Klueber/Desktop/University/Coursework/Fall 2019/GOV 1005 Data/Final project/Final project/Spending/2019 Spending split.xlsx")


#D.2 Write file to clean-data folder 

write_rds(Endowment_split, "clean-data/Endowment_split.rds", compress = "none")
```

```{r}
#E. Schools - student population and estimation of operating revenue


# E.1 Creating data table with number of students per school per year acc. to websites (NOTE: Radcliffe without real independent student population, FAS includes College & GSAS & SEAS) and Operating Revenue if schools published this in an annual report, otherwise 0 for now 

students <- data.frame("School" = c("FAS (College & GSAS & SEAS)", "Divinity_School", "Law_School", "Design_School", "Medical_School", "Kennedy_School", "Education_School", "Dental_School", "Business_School", "Public_Health_School"), "Student_pop" = c(11764, 268, 1990, 905, 1601, 1213, 876, 280, 1859, 1140), "Operating_revenue" = c(1571900000, 0, 0, 64500000, 703405004, 0, 0, 0, 766000000,0))


# E.2 For those not available: operating revenue determined by share of student population (total operating revenue from fy19_harvard_financial_report)

allocated <- students$Operating_revenue %>%
  sum()

  Delta = 5510570000 - allocated

students_est <- students %>%
  filter(Operating_revenue == 0) %>% 
  mutate(Student_pop_share = Student_pop/sum(Student_pop),
         Operating_revenue_est = Student_pop_share*Delta)
  

#E.3 Write files to clean-data folder 

write_rds(students, "clean-data/students.rds", compress = "none")
write_rds(students_est, "clean-data/students_est.rds", compress = "none")
```
