==> 2 Data prep_consolidating data.Rmd <==
---
title: "Emissions"
author: "Alexander Klueber"
date: "11/11/2019"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```
```{r}
library(janitor)
library(broom)
library(moderndive)
library(maps)
library(readxl)
library(tibble)
library(lubridate)
library(tidyverse)
```

```{r}
#A. Consolidation of carbon footprints, SEC filings and Bloomberg market capitalization into one central table

#A.1 Load relevant datasets

SEC_filings_cons <- read_rds("clean-data/SEC_filings_cons.rds")
CCF_combined <- read_rds("clean-data/ccf_combined.rds")
Market_caps <- read_xlsx("C:/Users/Alexander Klueber/Desktop/University/Coursework/Fall 2019/GOV 1005 Data/Final project/Final project/Funds/Market caps.xlsx")


#A.2 Join tables on Tickers and ISIN

HMC_CCF <- CCF_combined %>%
           separate(Ticker, c("Ticker", "Country_code")) %>% # Isolate matchable Ticker
           left_join(SEC_filings_cons, by = c("Ticker", "Ticker")) %>% # Join CCF with SEC_filings
           filter(!is.na(`Share of portfolio`)) %>% # Limiting to the observations that could be matched 
          left_join(Market_caps, by = c("ISIN", "ISIN"))%>%
           filter(MarketCap_09_30 != "N/A")
  
  #Recasting Market Capitalization as a number
  
  HMC_CCF$MarketCap_09_30 <- as.numeric(HMC_CCF$MarketCap_09_30)

  
#A.3 To rebase ownership share - calcualte share of portfolio available in data per year  

  HMC_CCF_year <- HMC_CCF %>%
                group_by(Year) %>%
                summarize(Total_share_year = sum(`Share of portfolio`))

  
#A.4 Rebase ownership share acc. to share of companies remaining in the dataset

HMC_CCF_v2 <- HMC_CCF %>%
              left_join(HMC_CCF_year, by = c("Year", "Year")) %>%
              mutate(Rebased_portfolio_share = `Share of portfolio`/Total_share_year)

#A.5 CONTROL: Share of portfolio available in data per year must all be 1

HMC_CCF_v2 %>%
  group_by(Year) %>%
  summarize(sum(Rebased_portfolio_share))


#A. 6 Take share of porfolio times 40.9 billion USD (acc. to Harvard Magazine: https://www.harvardmagazine.com/2019/09/harvard-endowment-increases-4-3-percent-to-40-9-billion) to extend available information as proxy to the entire . Also translate market cap from in millions to simple Euro amounts. Also calculate Harvard's ownership share in company. Also calculate the CO2 footprint in MT owned by Harvard 

HMC_CCF_v4 <- HMC_CCF_v2 %>%
  mutate(Ownership_USD = Rebased_portfolio_share*40900000000) %>% #Calculate Harvard's ownership stake of company as product of rebased portfolio share times 40.9 bn
  mutate(MarketCap_USD = MarketCap_09_30*1000000) %>% # Translate company market capitalization into simple dollars
  mutate(Ownership_share = Ownership_USD/MarketCap_USD) %>% # Calculate Harvard's ownership stake in companies
  mutate(Carbon_footprint_owned_MT = Ownership_share*Total) # Calculate Harvard's owned corporate carbon footprint as a result of that


#CHALLENGE: Dissonance between scaling and company values in years with no/little data. For 2005 2006, 2007 only First Energy Cooperation and SEARS Canada of the HMC portfolio provided CDP  data, therefore rebased share of portfolio is extremely high and ownership share exceeds 1. To resolve, I removed all years with scarce data by only looking at years where we had carbon footptint data for at least 30% of the portfolio ( from 2014 on):

HMC_CCF_v5 <- HMC_CCF_v4 %>%
  filter(Year > 2013) %>%
  select(Organisation, Year, Account_Number, Country, Industry, Ticker, Scope_1, Scope_2, Bio_CO2, Total, Rebased_portfolio_share, Ownership_share, Carbon_footprint_owned_MT) # Select relevant columns only

#A. 7 Write dataset to clean-data file as back-up

write_rds(HMC_CCF_v5, "clean-data/HMC_Carbon_Footprint.rds", compress = "none")

```

```{r}
#B. Consolidation of endowment split, Harvard Schools and carbon footprint data for a table of carbon footprint per student per school for 2014 - 2017

#(Note: endowment spending entirely on financing operations: "The endowment distributed 1.9 bn USD in the fiscal year ending June 30, 2019, contributing over a third of Harvard’s total operating revenue in that year" https://www.harvard.edu/about-harvard/harvard-glance/endowment / fy 2019 financial report)


#B. 1 Read in endowment spending data, and data on Harvard Schools 

Endowment_split <- read_rds("clean-data/Endowment_split.rds")
students <- read_rds("clean-data/students.rds")
students_est <- read_rds("clean-data/students_est.rds.")


#B. 2 Consolidate tables on spending data and Harvard Schools 

students_all <- students %>%
  left_join(students_est, by = c("School", "School")) %>%
  mutate(Operating_revenue_total = ifelse(Operating_revenue.x != 0, Operating_revenue.x, Operating_revenue_est))%>% # Combine data on available and estimated operating revenue
  select(School,Student_pop.x,Operating_revenue_total)%>%
  left_join(Endowment_split, by = c("School", "School"))%>%
  mutate(Endowment_financing=`Endowment income made available for operations`*Operating_revenue_total,
         Endowment_share_per_school = Endowment_financing/sum(Endowment_financing)) #Using endowment income made available for operations from FY 2019 Harvard annual report and total operating revenue to determine an operational share of endowment (not owned by the schools as assets, but used by the schools to finance their operations aka the experience of the student attending)

schools <- students_all %>%
  select(School, Share_carbon_ownership = Endowment_share_per_school, Student_pop.x)

write_rds(schools, "clean-data/Schools_endowment payout.rds", compress = "none")

#B. 3 Calculate total Carbon footprint of Harvard per year

CCF_per_year <- HMC_CCF_v5 %>%
  group_by(Year)%>%
  summarize(Carbon_footprint_owned_MT_year = sum(Carbon_footprint_owned_MT)) %>%
  spread(key = Year, value = Carbon_footprint_owned_MT_year)

#B. 4 Calculate total Carbon footprint of Harvard per school and per student using loops

  CCF_per_school <- data.frame()
  for (j in 1:4){
    for(i in 1:10){
    CCF_per_school[i,j] <- CCF_per_year[1,j]*students_all[i,6]
    }
  }
  
  CCF_per_student <- data.frame()
  for (j in 1:4){
    for(i in 1:10){
    CCF_per_student[i,j] <- CCF_per_school[i,j]/students_all[i,2]
    }
  }

#B. 5 Add a column with school names and appropriate headers for school data
      
  CCF_per_school <- add_column(CCF_per_school, students_all$School, .before = 1)
  colnames(CCF_per_school) <- c("Schools",2014,2015,2016,2017)

#B. 6 Add a total row and gather data across all years in one column, divide carbon footprint by 1000 to make it easier to digest and year numeric to allow for limits in X axis during graphing

    CCF_per_school <- CCF_per_school %>%
  adorn_totals("row")%>%
    gather(key = "Year", value = "Carbon_footprint_in_MT", 2:5)%>%
    mutate(Carbon_footprint_in_MT= (Carbon_footprint_in_MT/1000),
           Year =as.numeric(Year))

#B. 7 Write data to clean-data folder for later utilitization
    
  write_rds(CCF_per_school, "clean-data/CCF_per_school.rds", compress = "none")
  
#B. 8 Add a column with school names and appropriate headers for student data
    
  CCF_per_student <- add_column(CCF_per_student, students_all$School, .before = 1)
  colnames(CCF_per_student) <- c("Schools",2014,2015,2016,2017)
  
# B. 9 Filter for 2017 data and Introduce average carbon footprint and various carbon footprint reduction levers
    # Average_carbon_footprint_US of 15.8 according to monthly Energy Review November 2019 by the EPA (Table 1.7: Primary Energy Consumption, Energy Expenditures, and Carbon Dioxide Emissions Indicators) Emissions per Capita column for 2017 
    # Reduction on the individual level per measure from Wynes & Nicholas (2017): The climate mitigation gap
  
  CCF_per_student_red <- CCF_per_student %>%
    select("Schools", "2017") %>%
    mutate(`Average annual carbon footprint in US per individual` = 15.8,
           `Additional carbon footprint via Harvard Endowment` = round(`2017`,2),
           `Reallocate Harvard endowment to assets with zero emissions` = -round(`2017`,2),
           `Live car free` = -3.08,
           `Avoid a transatlantic flight` = -1.6,
           `Buy green energy` = -1.4,
           `Renounce all meat` = -0.91) %>%
          select(-`2017`) %>%
    gather(key = "Type", value = "Carbon_footprint_in_MT", 2:8) %>% # Gather all levers/footprints in one column
    mutate(Action = ifelse(Carbon_footprint_in_MT > 0, "Carbon footprint", "Reduction measure"))
  
#B.10 Locking in order of types as to be displayed in the bar chart
  
  CCF_per_student_red$Type <- factor(CCF_per_student_red$Type, rev(unique(CCF_per_student_red$Type)))

#B. 11 Write data to clean-data folder for later utilitization
  
  write_rds(CCF_per_student_red, "clean-data/CCF_per_student_red.rds", compress = "none")

```
```{r}
# C. 1 Reduce full dataset to 2017 and Organisation, Ownership share, Carbon_footprint_owned_MT

Emitter_table <- HMC_CCF_v5 %>%
                 filter(Year == 2017) %>%
                 mutate(Ownership_share = paste(round(Ownership_share*100,2),"%",sep=""),
                        Carbon_footprint_owned_MT = round(Carbon_footprint_owned_MT,0))%>%
                 select(`Portfolio company`= Organisation, 
                        `Harvard share in ownership (scaled)` = Ownership_share, 
                        `MT inherited in 2017` = Carbon_footprint_owned_MT) %>%
                 arrange(desc(`MT inherited in 2017`)) 

#C. 2 Write data to clean-data folder for later utilitization
                 
write_rds(Emitter_table, "clean-data/Emitter_table.rds", compress = "none")

```

```{r}
# D. Replicate a random average for CDP companies for 2014 - 2017 through 1000 times randomly selecting 562 companies from the available sample for each year of the years and compare average of these to the average we saw in our real data.

# D. 1 Create mean for HMC portfolio companies for each of the years

CCF_per_year_average_HMC <- HMC_CCF_v5 %>%
  group_by(Year)%>%
  summarize(Average_carbon_footprint_per_company = mean(Total)/1000)

# D. 2 Repeat 1000 times random selection of size 562 and create average for each of the years

CCF_per_year_average_all_2014 <- CCF_combined %>%
  filter(Year==2014) %>%
  rep_sample_n(size = 562, replace = TRUE, reps = 1000) %>%
  group_by(replicate) %>%
  summarize(Average_carbon_footprint_per_company = mean(Total)/1000)%>%
  mutate(Year = 2014)

CCF_per_year_average_all_2015 <- CCF_combined %>%
  filter(Year==2015) %>%
  rep_sample_n(size = 562, replace = TRUE, reps = 1000) %>%
  group_by(replicate) %>%
  summarize(Average_carbon_footprint_per_company = mean(Total)/1000)%>%
  mutate(Year = 2015)

CCF_per_year_average_all_2016 <- CCF_combined %>%
  filter(Year==2016) %>%
  rep_sample_n(size = 562, replace = TRUE, reps = 1000) %>%
  group_by(replicate) %>%
  summarize(Average_carbon_footprint_per_company = mean(Total)/1000)%>%
  mutate(Year = 2016)

CCF_per_year_average_all_2017 <- CCF_combined %>%
  filter(Year==2017) %>%
  rep_sample_n(size = 562, replace = TRUE, reps = 1000) %>%
  group_by(replicate) %>%
  summarize(Average_carbon_footprint_per_company = mean(Total)/1000)%>%
  mutate(Year = 2017)

# D.3 Combine all years together into one table

CCF_per_year_average_all <- rbind(CCF_per_year_average_all_2014,CCF_per_year_average_all_2015, CCF_per_year_average_all_2016,CCF_per_year_average_all_2017)

# D.4 Write files to clean-data for later visualization

write_rds(CCF_per_year_average_HMC, "clean-data/CCF_per_year_average_HMC.rds", compress = "none")
write_rds(CCF_per_year_average_all, "clean-data/CCF_per_year_average_all.rds", compress = "none")

```
